CLDR-14925 Preliminary organizing for snapshot of Priority Items Summary
-Move Dashboard-specific code in VettingViewerQueue.java to new Dashboard.java

-Move STUsersChoice into new STUsersChoice.java

-Move getChoiceSetForOrg into VettingViewer.java
